Bugfix in kernel, update to milestone description, and deleted leftover unnecessary files
The kernel code had a bug where it was possible to create mappings that are
larger than the region covered by the provided frame capability.  This commit
fixes this and also removes a number of unnecessary files from the tree.

Further the milestone description for milestone 1 has been updated to give
some information about the `mapping` parameter to vnode_map().